Fix : 리팩토링

+++ back/src/services/authService.ts
@@ -7,14 +7,12 @@ import { successApiResponseDTO } from '../utils/successResult';
 import { userCalculatePageInfo } from '../utils/pageInfo';
 import { PaginationResponseDTO } from '../dtos/diaryDTO';
 import { emailToken, sendEmail } from '../utils/email';
+import { emptyApiResponseDTO } from '../utils/emptyResult';
+import { getMyWholeFriends } from './friendService';
 
 const prisma = new PrismaClient();
 
-export const createUser = async (inputData: {
-  username: string;
-  password: string;
-  email: string;
-}) => {
+export const createUser = async (inputData: Prisma.UserCreateInput) => {
   const { username, password, email } = inputData;
 
   // 비밀번호를 해시하여 저장 (안전한 비밀번호 저장)
@@ -56,11 +54,9 @@ export const getAllUsers = async (
   page: number,
   limit: number,
 ) => {
-  const pageSize = 10;
-
   const userList = await prisma.user.findMany({
-    take: pageSize,
-    skip: (page - 1) * pageSize,
+    take: limit,
+    skip: (page - 1) * limit,
     orderBy: {
       createdAt: 'asc',
     },
@@ -82,7 +78,7 @@ export const getAllUsers = async (
       },
     },
   });
-
+  console.log(userList);
   userList.forEach((user) => {
     user.isFriend = user.friendS.length > 0 || user.friendR.length > 0;
 
@@ -116,63 +112,32 @@ export const getMyFriends = async (
   page: number,
   limit: number,
 ) => {
-  const pageSize = 10;
+  const startTime = new Date().getTime();
 
-  const userList = await prisma.user.findMany({
-    take: pageSize,
-    skip: (page - 1) * pageSize,
-    orderBy: {
-      createdAt: 'asc',
-    },
-    include: {
-      friendS: {
-        where: {
-          status: true,
-        },
-      },
-      friendR: {
-        where: {
-          status: true,
-        },
-      },
-      Diary: {
-        where: {
-          authorId: userId,
-        },
-        orderBy: {
-          createdAt: 'desc',
-        },
-        take: 1,
+  const friendList = await getMyWholeFriends(userId);
+  console.log(friendList);
+  const friendIds = friendList.map((friend) => friend.sentUserId);
+  console.log(friendIds);
+  const friendsInfo = await prisma.user.findMany({
+    take: limit,
+    skip: (page - 1) * limit,
+    where: {
+      id: {
+        in: friendIds, // 친구의 ID 목록
       },
     },
   });
+  console.log(friendsInfo) 
 
-  const filteredUsers = userList.filter((user) => {
-    const areFriends = user.friendS.length > 0 || user.friendR.length > 0;
-
-    if (user.id !== userId && areFriends) {
-      // 자기 자신이 아니면서 친구인 경우만 결과에 포함
-      return true;
-    }
-
-    return false;
-  });
-
-  filteredUsers.forEach((user) => {
-    user.isFriend = user.friendS.length > 0 || user.friendR.length > 0;
-
-    if (user.Diary.length > 0) {
-      user.latestEmoji = user.Diary[0].emoji;
-    } else {
-      user.latestEmoji = '❎';
-    }
-  });
-  const totalItem = filteredUsers.length;
+  const endTime = new Date().getTime();
+  const responseTime = endTime - startTime;
+  console.log(responseTime);
+  const totalItem = friendsInfo.length;
   const totalPage = Math.ceil(totalItem / limit);
 
   const pageInfo = { totalItem, totalPage, currentPage: page, limit };
 
-  const userResponseDataList = filteredUsers.map((user) =>
+  const userResponseDataList = friendsInfo.map((user) =>
     plainToClass(userResponseDTO, user, { excludeExtraneousValues: true }),
   );
 
@@ -200,6 +165,14 @@ export const getUserInfo = async (userId: string) => {
   return response;
 };
 
+export const logout = async(userId : string) => {
+  await prisma.refreshToken.deleteMany({
+    where: {
+      userId: userId,
+    },
+  });
+}
+
 export const updateUserService = async (
   userId: string,
   inputData: Prisma.UserUpdateInput,
@@ -221,11 +194,24 @@ export const updateUserService = async (
       },
     },
   });
-  const response = successApiResponseDTO(updatedUser);
+  const UserResponseDTO = plainToClass(userResponseDTO, updatedUser, {
+    excludeExtraneousValues: true,
+  });
+  const response = successApiResponseDTO(UserResponseDTO);
   return response;
 };
 
 export const deleteUserService = async (userId: string) => {
+  const user = await prisma.user.findUnique({
+    where: { id: userId },
+  });
+
+  if (!user) {
+    // 사용자를 찾을 수 없는 경우 적절한 오류 처리를 수행
+    const response = emptyApiResponseDTO();
+    return response;
+  }
+
   // 사용자의 refreshTokens 먼저 삭제
   await prisma.refreshToken.deleteMany({
     where: {
@@ -235,9 +221,9 @@ export const deleteUserService = async (userId: string) => {
 
   // 사용자의 친구 관계 삭제
   await prisma.friend.deleteMany({
-    where : {
+    where: {
       OR: [{ sentUserId: userId }, { receivedUserId: userId }],
-    }
+    },
   });
 
   // 사용자의 다이어리 삭제
@@ -253,11 +239,17 @@ export const deleteUserService = async (userId: string) => {
       id: userId,
     },
   });
-
-  return '사용자가 삭제되었습니다.';
 };
 
 export const forgotUserPassword = async (email: string) => {
+  // 데이터베이스에서 사용자 이메일로 사용자 조회
+  const user = await prisma.user.findUnique({ where: { email } });
+
+  if (!user) {
+    const response = emptyApiResponseDTO();
+    return response;
+  }
+
   // 임시 비밀번호 생성
   const tempPassword = generateRandomPassowrd();
   const saltRounds = 10;
@@ -281,6 +273,14 @@ export const forgotUserPassword = async (email: string) => {
 };
 
 export const resetUserPassword = async (email: string, password: string) => {
+    // 데이터베이스에서 사용자 이메일로 사용자 조회
+    const user = await prisma.user.findUnique({ where: { email } });
+
+    if (!user) {
+      const response = emptyApiResponseDTO();
+      return response;
+    }
+
   const saltRounds = 10;
 
   // 새로운 비밀번호를 해시하여 저장
@@ -360,7 +360,7 @@ export const getUsers = async (
   return response;
 };
 
-export const emailLinked = async(email : string) => {
+export const emailLinked = async (email: string) => {
   const user = await prisma.user.create({
     data: {
       email,
@@ -396,9 +396,9 @@ export const emailLinked = async(email : string) => {
         <p><a href = "${verifyUrl}">Verify Email</a></p>
         <p>${result.expires}</p>`,
   );
-}
+};
 
-export const verifyToken = async (token : string) => {
+export const verifyToken = async (token: string) => {
   const user = await prisma.user.findFirst({
     where: {
       verificationToken: token,
@@ -409,7 +409,7 @@ export const verifyToken = async (token : string) => {
   });
 
   if (!user) {
-    throw ({ message: '토큰이 유효하지 않습니다.' });
+    throw { message: '토큰이 유효하지 않습니다.' };
   }
 
   await prisma.user.update({
@@ -422,15 +422,19 @@ export const verifyToken = async (token : string) => {
       verificationTokenExpires: null,
     },
   });
-}
+};
 
-export const registerUser = async(email : string, username : string, password : string) => {
+export const registerUser = async (
+  email: string,
+  username: string,
+  password: string,
+) => {
   const user = await prisma.user.findUnique({
     where: { email },
   });
 
   if (!user || !user.isVerified) {
-    throw ({ message: '이메일 인증이 필요합니다.' });
+    throw { message: '이메일 인증이 필요합니다.' };
   }
 
   // 비밀번호를 해시하여 저장 (안전한 비밀번호 저장)
@@ -440,7 +444,7 @@ export const registerUser = async(email : string, username : string, password : 
     where: { id: user.id },
     data: {
       username,
-      password : hashedPassword,
+      password: hashedPassword,
     },
   });
 
@@ -450,4 +454,4 @@ export const registerUser = async(email : string, username : string, password : 
 
   const response = successApiResponseDTO(UserResponseDTO);
   return response;
-}
+};
